Numerous fixes
Try to correct the hacky way we try to determine that we're
setting the carry flag in VEX. I knew that it was UB to
count on the order of functions in the object file matching
the source file, bit in this case it doesn't.
Maybe I should try to set a global variable.

Make sigframe code more like amd64, probably redundant. Add
tracking for modified registers.

Use fp_reg for floating point register struct.

Remove SYS_freebsd10_pipe, it's not implemented on arm64.

Add clobbers for memcheck leak tests.