Fix signal mask issues
I'd copied stuff from the amd64 vg_sigframe including the return
address and phandler. That meant that ucontext wasn't at the
start of the struct. PRE(sys_fake_sigreturn) gets passed a
pointer to vg_sigframe as its first argument (from the stack
pointer sp). Result: uc, the pointer to the ucontext was out
by 16 bytes. That caused the guest signal masks to be
filled with rubbish.
it didn't affect very many tescases